+++ hyperdbg/include/SDK/Imports/HyperDbgVmmImports.h
@@ -185,6 +185,9 @@ VmFuncNmiBroadcastInvalidateEptAllContexts(UINT32 CoreId);
 IMPORT_EXPORT_VMM BOOLEAN
 VmFuncVmxGetCurrentExecutionMode();
 
+IMPORT_EXPORT_VMM INT32
+VmFuncVmxCompatibleStrcmp(const CHAR * Address1, const CHAR * Address2);
+
 //////////////////////////////////////////////////
 //            Configuration Functions 	   		//
 //////////////////////////////////////////////////
@@ -330,6 +333,9 @@ CheckAddressValidityUsingTsx(CHAR * Address);
 
 IMPORT_EXPORT_VMM BOOLEAN
 CheckAccessValidityAndSafety(UINT64 TargetAddress, UINT32 Size);
+
+IMPORT_EXPORT_VMM BOOLEAN
+CheckAddressPhysical(UINT64 PAddr);
 
 IMPORT_EXPORT_VMM UINT32
 CheckAddressMaximumInstructionLength(PVOID Address);